fix: 댓글 컨트롤러 수정

+++ src/main/java/com/kb04/starroad/Controller/CommentController.java
@@ -144,16 +144,26 @@ public class CommentController {
             @RequestParam("no") int commentNo,
             @ApiIgnore HttpSession session) {
 
+
+        ModelAndView mav = new ModelAndView();
         if (session.getAttribute("currentUser") == null) {
-            ModelAndView mav = new ModelAndView("redirect:/starroad/login");
+            mav = new ModelAndView("redirect:/starroad/login");
             mav.addObject("message", "로그인 후에 댓글을 작성할 수 있습니다.");
             return mav;
         }
 
-        int commentDelete = commentService.findByNo(commentNo).get().getBoard().getNo();
-        commentService.deleteComment(commentNo);
+        MemberDto currentUser = (MemberDto) session.getAttribute("currentUser");
+        String currentUserId = currentUser.getId();
+        if (commentService.canDelete(commentNo, currentUserId)) {
+            // 현재 사용자가 게시글 삭제 가능한 경우
+            int boardNo = commentService.deleteComment(commentNo);
+            mav.setViewName("redirect:/starroad/board/detail?no=" + boardNo);
 
-        ModelAndView mav = new ModelAndView("redirect:/starroad/board/detail?no=" + commentDelete);
+        } else {
+            // 현재 사용자가 게시글 삭제 불가
+            // 능한 경우 or 게시글 존재하지 않는 경우
+            mav.setViewName("board/deleteError");
+        }
         return mav;
     }
